requirements.txt
streamlit
streamlit-chat
pandas
python-dotenv
groq
